qiwitest/requirements.txt
playwright==1.40.0
pytest==7.4.4
pytest-playwright==0.4.1
python-dotenv==1.0.1